Bundle several eth-testing mock setups into a `mockMainnet` method
