+++ packages/safe-tools-client/tests/helpers/index.ts
@@ -1,6 +1,8 @@
+import { ERC20ABI } from '@cardstack/cardpay-sdk';
 import { truncateMiddle } from '@cardstack/ember-shared/helpers/truncate-middle';
 import { MockLocalStorage } from '@cardstack/safe-tools-client/utils/browser-mocks';
 import { click, TestContext, visit } from '@ember/test-helpers';
+import IUniswapV2Pair from '@uniswap/v2-core/build/IUniswapV2Pair.json';
 import {
   setupApplicationTest as upstreamSetupApplicationTest,
   setupRenderingTest as upstreamSetupRenderingTest,
@@ -9,7 +11,6 @@ import {
 import { generateTestingUtils } from 'eth-testing';
 import { MetaMaskProvider } from 'eth-testing/lib/providers';
 import { TestingUtils } from 'eth-testing/lib/testing-utils';
-
 // This file exists to provide wrappers around ember-qunit's / ember-mocha's
 // test setup functions. This way, you can easily extend the setup that is
 // needed per test type.
@@ -22,13 +23,20 @@ export const TEST_ACCOUNT_1 = '0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266';
 // Hardhat default test account 2 (PK 0x59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d)
 export const TEST_ACCOUNT_2 = '0x70997970C51812dc3A010C7d01b50e0d17dc79C8';
 
+type CardstackTestingUtils = TestingUtils & {
+  mockMainnet(): void;
+};
 declare module '@ember/test-helpers' {
   interface TestContext {
-    mockWalletConnect: TestingUtils;
+    mockWalletConnect: CardstackTestingUtils;
     mockMetaMask: TestingUtils;
     mockLocalStorage: MockLocalStorage;
   }
 }
+
+const USDC_TOKEN_ADDRESS = '0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48';
+const WETH_TOKEN_ADDRESS = '0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2';
+const UNISWAP_V2_ADDRESS = '0xb4e16d0168e52d35cacd2c6185b44281ec28c9dc';
 
 function setupApplicationTest(hooks: NestedHooks, options?: SetupTestOptions) {
   upstreamSetupApplicationTest(hooks, options);
@@ -42,11 +50,42 @@ function setupApplicationTest(hooks: NestedHooks, options?: SetupTestOptions) {
     this.mockWalletConnect = generateTestingUtils({
       providerType: 'WalletConnect',
       verbose: true,
-    });
+    }) as CardstackTestingUtils;
+
+    this.mockWalletConnect.mockMainnet = () => {
+      this.mockWalletConnect.mockChainId(1);
+      const mockUsdcContract = this.mockWalletConnect.generateContractUtils(
+        ERC20ABI,
+        USDC_TOKEN_ADDRESS
+      );
+      mockUsdcContract.mockCall('decimals', ['6'], undefined, {
+        persistent: true,
+      });
+
+      const mockWethContract = this.mockWalletConnect.generateContractUtils(
+        ERC20ABI,
+        WETH_TOKEN_ADDRESS
+      );
+      mockWethContract.mockCall('decimals', ['18'], undefined, {
+        persistent: true,
+      });
+
+      const mockUniswapContract = this.mockWalletConnect.generateContractUtils(
+        IUniswapV2Pair.abi,
+        UNISWAP_V2_ADDRESS
+      );
+      mockUniswapContract.mockCall(
+        'getReserves',
+        ['1', '1', '10000'],
+        undefined,
+        {
+          persistent: true,
+        }
+      );
+    };
 
     this.mockWalletConnect.mockNotConnectedWallet();
     this.mockWalletConnect.mockAccounts([TEST_ACCOUNT_2]);
-    this.mockWalletConnect.mockChainId(1);
 
     this.owner.register(
       'ethereum-provider:wallet-connect',
